+++ app/src/main/java/eu/kanade/tachiyomi/data/database/models/TrackImpl.kt
@@ -14,7 +14,7 @@ class TrackImpl : Track {
 
     override lateinit var title: String
 
-    override var last_chapter_read: Int = 0
+    override var last_chapter_read: Float = 0F
 
     override var total_chapters: Int = 0
